Added a very basic sampling profiler
Added a sampling profiler with an interface that provides a function
aggregation. Testing indicates that it is fast enough to enable
unconditionally. I have a patch for Scribunto lined up that adds the
results to the limit report for slow pages (>1s).

Unfortunately it proved to be impossible to start and stop the timer
every time Lua was entered. With a typical test case, it would miss
about 50% of execution time and show a strong bias towards long-running
invocations, due to details of the latency between timer expiration and
signal delivery. So the timer will start the first time Lua is called,
and keep running until the LuaSandbox object is destroyed.

I made some efforts to provide a meaningful function name, slightly
beyond what Lua does by default.

I moved the signal handler deregistration to post-deactivate so that the
profiler would be stopped first.

Change-Id: I0b83a914ae993418b58a059003309f6c9bafd7de