Add LuaSandbox::wrapPhpFunction
Currently, the only way for PHP to pass a function to Lua is to pass
back a LuaSandboxFunction that it received from Lue. This means, for
example, that PHP cannot implement a Lua iterator function except by
registering a library holding the function or by using loadString.

This changeset adds LuaSandbox::wrapPhpFunction, which takes a PHP
callable and returns a LuaSandboxFunction wrapping it.

Change-Id: I2e552799dbea32879397fd98b6678abb854dacac