Zero timeout flag on unpause
When the normal limit timer fires while timers are paused, the timestamp
is recorded and the timer is recheduled on unpause. But we forgot to
zero out that timestamp, so the next time it is paused and unpaused it
would reschedule the timer again. This may lead to incorrect behavior.

Change-Id: Ib4118f7d1aed218ceb2fb567a65bb60a92ad21db