Allow memory over-allocation in unprotected Lua calls
If the configured memory limit happens to be hit in some of the setup
code rather than in user-provided Lua code itself (e.g. in the
lua_pushcclosure function needed to get the function on the stack before
we can use lua_pcall), we can wind up with a PHP fatal error rather than
the expected script error.

A simple solution to this problem is to allow the non-user-provided code
to use a little extra memory, so the error condition can happen in the
user-provided Lua code instead.

Bug: 59130
Change-Id: I081a05daa5501238ef7b54a14a5cdb5a8fafce6f