Fix getProfilerFunctionReport
zend_hash_get_current_data_ex is effectively an accessor for the pData
pointer of the hash element, so the pointer returned still needs to be
dereferenced to get the actual value stored. As-is, we're using the
pointer itself cast to a size_t as the time for each function.

Change-Id: I430121166d3989340697188c52cdbb072c519665