Dynamically allocate luasandbox_timer instances
With the change of the limit timer to SIGEV_THREAD, the old value of
LUASANDBOX_MAX_TIMERS isn't enough to run Scribunto's unit tests anymore
if PHPUnit decides to create all the test classes at once. Rather than
keep bumping the limit, let's change it to dynamically allocate the
needed data structures.

To do this, we implement a simple open-address hash table with an rwlock
to make accesses thread-safe.

Change-Id: I6d85b746c8e375b67d340be404fd7a94c5963c65