Build for both PHP<< 5.5 and PHP and HHVM
Allow on build time detection of HHVM existence, allowing the luasandbox
to be built for non HHVM-enabled distributions as well. Achieve this by
detecting HHVM via the presence of hphpize binary and define the already
split split in this patch makefile's targets accordingly. In order to achieve
Debian package build dependency satisfaction, use boolean algebra rules
to break down the:
	PHP5.3 or (HHVM - and dependenies - and PHP5.5)
expression to the equivalent ones. The | (or) operator takes precedence
when calculating dependencies so this installs either PHP5 versions less
than 5.5 (precise has 5.3) or PHP5 5.5 and HHVM and dependencies.

Bug: T88798

Change-Id: I445fc75f66c470da5dbe32fd8a4f8c2c7606bada