Don't try to automatically get the version from the debian/changelog
The debian/ files are being moved to a separate branch to make it easier
for actual Debian packaging as recommended in
<https://wiki.debian.org/UpstreamGuide#Pristine_Upstream_Source>. Since
the debian/changelog file is no longer accessible, set
LUASANDBOX_VERSION to an alpha version. When a new version is released
and tagged, it should be updated as well.

Change-Id: I070266c9c2adbf78203c961d3226e477197c7a15