Fix getProfilerFunctionReport sorting in HHVM
HHVM has a bug in their zend_hash_sort() where it ignores the passed
'compar' function, instead always doing a ksort. This breaks
getProfilerFunctionReport(). It's said to be unlikely that HHVM would
care enough to fix this.

On the plus side, HHVM has ext_luasandbox.php where we can put actual
PHP code to wrap the method and call arsort() to do the sorting.

Bug: T158029
Change-Id: I0a6d60fd981ab41f8b39b93f0164ee35214e6147