Redo PHP->Lua data structure recursion protection
We've been getting a fairly significant level of false positives from
the recursion protection in WMF production. One theory as to the cause
of this is the following sequence of events:
* Something runs into a Lua error while converting a nested data
  structure from PHP to Lua. Probably the memory limit.
* Lua jumps via exception out of luasandbox_push_hashtable, without
  calling luasandbox_unprotect_recursion for the arrays/objects in the
  process of being converted. Thus, their addresses are left in
  protectionSet.
* Later, some other call during another request happens to use the same
  address for a different array/object. Since the address is still in
  protectionSet, it gets reported as "recursion detected".

This patch changes the recursion detection to use a locally-allocated
hash table instead of a global set, like the Lua->PHP code path does. It
also operates on the zvals rather than the HashTables because it turns
out HHVM's Z_OBJPROP_P() returns a different HashTable on every call,
even for the same object, while zval pointer identity seems to work
fine.

While the Zend versions of this code don't have the same issue as HHVM,
that's mainly because there the recursed flag is stored in the HashTable
instance itself so the garbage doesn't hang around like a land mine for
some future request. It's still probably leaving nApplyCount with
garbage that might break something somewhere before the zval in question
goes away.

Bug: T166348
Change-Id: I745a77011a1e134ffa4f19db89840f1cf9c6e53c