Catch Lua errors at PHP→Lua call boundaries
I081a05daa tried to avoid memory errors in unprotected code by allowing
unprotected code to allocate a bit beyond the normal memory limit.

But it turns out that's not sufficient, since in a Lua→PHP→Lua call the
setup for the second Lua bit should also be considered "unprotected",
but at that point it's easy enough for the data being passed from PHP to
Lua to exceed the "slop" space.

But even that too wouldn't be sufficient, because Lua errors other than
memory errors should also not be thrown across that boundary. Attempting
to do so results in HHVM catching it and reporting an "unknown
exception" fatal error, and in Zend PHP5 tends to result in a segfault
(in limited testing PHP7 seems fine, although I have no idea why).

So, to really fix this, we need to catch these Lua errors at each
PHP→Lua call boundary. Lua provides a lua_cpcall() method for this
purpose, although actually making use of it is a bit of a pain as it
requires a callback function and passing any necessary parameters
(including the tsrm_ls parameter hidden behind the TSRMLS_* macros) to
it via a struct that's sent as a light userdata on the Lua stack.

Reviewing the methods provided by this extension, the following don't
make calls into Lua and so are safe:
* All the exception classes
* LuaSandbox::getVersionInfo()
* LuaSandbox::setMemoryLimit()
* LuaSandbox::getMemoryUsage()
* LuaSandbox::getPeakMemoryUsage()
* LuaSandbox::setCPULimit()
* LuaSandbox::getCPUUsage()
* LuaSandbox::pauseUsageTimer()
* LuaSandbox::unpauseUsageTimer()
* LuaSandbox::enableProfiler()
* LuaSandbox::disableProfiler()
* LuaSandbox::getProfilerFunctionReport()
* LuaSandboxFunction::__construct()

The following all needed protection added:
* LuaSandbox::loadString()
* LuaSandbox::loadBinary()
* LuaSandbox::callFunction()
* LuaSandbox::wrapPhpFunction()
* LuaSandbox::registerLibrary()
* LuaSandboxFunction::call()
* LuaSandboxFunction::dump()

Since the allocator code from I081a05daa isn't serving the purpose, I'm
removing it in this patch. I'm also fixing a minor bug in the allocator
in that calls with nsize < osize (i.e. reducing memory usage) should
always be allowed regardless of the memory usage status, so that if Lua
tries to clean up after memory usage is exceeded it can actually do so.

Bug: T171392
Change-Id: Iaa88053183d51a698750c6df4b40ba7afe00c913