Fix PHP 7 object layout
PHP 7 uses the "struct hack" in zend_object, so it needs to be at the
end of the structure rather than the start and the allocation needs to
take the struct hack's size into account.

This also means a simple cast doesn't work anymore. Most of the time
we're already using a macro for casts from a zval (rather than having a
zend_object* to cast), so there's little to clean up there.

Bug: T205370
Change-Id: Icde045fc86f0f5e68a42884a192e3cd3850e96a2