Remove memory leaks in data_conversion.c
The variable value was not released in luasandbox_lua_pair_to_array
in data_conversion.c for non-string and non-integer keys and a collision
with an integer-like string.

This caused the second test, tests/array-key-conversion.phpt to fail.

By inserting zval_ptr_dtor(&value); in three places, the memory leak is removed.

Bug: T263129.
Change-Id: I64901c60b917e9e97be539ca19a0f9f283cf0ab8